Section 2 Update:
Tony and I have been working on the catipilar segment which will be 5 1x1 moving platforms that move very similarly to the game of snake. We are plotting points for the head and then will attach the body there after. The lead up of our section to the catipilar is complete minus operable moving platforms.