Updated Yoki Gym
Updated the Yoki gym and speed / jump height parameters of the Yoki to be in line with specifications

The feel is much more Mario like now. The Yoki moves slower but is still a good bit faster than the Girl and the jump can just barely make 7 units high, but the jump needs perfect execution. 6 units is cleared relatively easily. Give it a test and let me know what you guys thing, but it feels "right" to me where its at.