Bug Fixed
The player doesn't stick to ceilings anymore. In the future, if and jump scripts are edited or created, make sure that an "isGrounded" boolean is included in the selection statement. I also edited some trivial things about the Player prefab, like making the capsule object invisible, and adding controls for using the spacebar to jump.